+++ Realm/RLMRealm.mm
@@ -138,6 +138,7 @@ NSData *RLMRealmValidatedEncryptionKey(NSData *key) {
 
     RLMCheckForUpdates();
     RLMSendAnalytics();
+    RLMWorkaroundRadar31252694();
 }
 
 - (instancetype)initPrivate {